Slim down braintrust core python library.
We pare down the braintrust core python library to just those things
shared between `sdk` and `autoevals`, and move the rest into `sdk`. We
also maintain a version for the python core separately from the
javascript core.

This will simplify package deployment as we should generally just update
the javascript core library.
