+++ core/py/src/braintrust_core/score.py
@@ -1,8 +1,9 @@
 import dataclasses
+import sys
 from abc import ABC, abstractmethod
 from typing import Dict, Optional
 
-from .util import SerializableDataClass, eprint
+from .serializable_data_class import SerializableDataClass
 
 
 @dataclasses.dataclass
@@ -32,8 +33,9 @@ class Score(SerializableDataClass):
         if self.score is not None and (self.score < 0 or self.score > 1):
             raise ValueError(f"score ({self.score}) must be between 0 and 1")
         if self.error is not None:
-            eprint(
-                "The error field is deprecated, as errors are now propagated to the caller. The field will be removed in a future version of the library"
+            print(
+                "The error field is deprecated, as errors are now propagated to the caller. The field will be removed in a future version of the library",
+                sys.stderr,
             )
