+++ core/py/src/braintrust_core/serializable_data_class.py
@@ -1,14 +1,6 @@
 import dataclasses
 import json
-import sys
-import urllib.parse
-from typing import Dict, List, Set, Tuple, Union, get_origin
-
-
-# Taken from
-# https://stackoverflow.com/questions/5574702/how-do-i-print-to-stderr-in-python.
-def eprint(*args, **kwargs):
-    print(*args, file=sys.stderr, **kwargs)
+from typing import Dict, Union, get_origin
 
 
 class SerializableDataClass:
@@ -65,53 +57,3 @@ class SerializableDataClass:
             else:
                 filtered[k] = v
         return cls(**filtered)
-
-
-def coalesce(*args):
-    """Returns the first non-None value in the list of `args`, or `None` if they
-    are all `None`.
-    """
-
-    for a in args:
-        if a is not None:
-            return a
-    return None
-
-
-def merge_dicts_with_paths(merge_into: Dict, merge_from: Dict, path: Tuple[str], merge_paths: Set[Tuple[str]]):
-    """Merges merge_from into merge_into, destructively updating merge_into. Does not merge any further than
-    merge_paths."""
-
-    if not isinstance(merge_into, dict):
-        raise ValueError("merge_into must be a dictionary")
-    if not isinstance(merge_from, dict):
-        raise ValueError("merge_from must be a dictionary")
-
-    for k, merge_from_v in merge_from.items():
-        full_path = path + (k,)
-        merge_into_v = merge_into.get(k)
-        if isinstance(merge_into_v, dict) and isinstance(merge_from_v, dict) and full_path not in merge_paths:
-            merge_dicts_with_paths(merge_into_v, merge_from_v, full_path, merge_paths)
-        else:
-            merge_into[k] = merge_from_v
-
-    return merge_into
-
-
-def merge_dicts(merge_into: Dict, merge_from: Dict):
-    """Merges merge_from into merge_into, destructively updating merge_into."""
-
-    return merge_dicts_with_paths(merge_into, merge_from, (), set())
-
-
-def encode_uri_component(name):
-    """Encode a single component of a URI. Slashes are encoded as well, so this
-    should not be used for multiple slash-separated URI components."""
-
-    return urllib.parse.quote(name, safe="")
-
-
-def _urljoin(*parts):
-    return "/".join(
-        p for p in [x.strip("/") if i < len(parts) - 1 else x.lstrip("/") for i, x in enumerate(parts)] if p.strip()
-    )
