+++ core/py/src/braintrust_core/version.py
@@ -1,1 +1,1 @@
-VERSION = "0.0.52"
+VERSION = "0.0.53"
