+++ py/src/braintrust/gitutil.py
@@ -6,7 +6,7 @@ import threading
 from functools import lru_cache as _cache
 from typing import List, Optional
 
-from braintrust_core.git_fields import GitMetadataSettings, RepoInfo
+from .git_fields import GitMetadataSettings, RepoInfo
 
 # https://stackoverflow.com/questions/48399498/git-executable-not-found-in-python
 os.environ["GIT_PYTHON_REFRESH"] = "quiet"
